+++ cyberdrop_dl/crawlers/__init__.py
@@ -82,8 +82,8 @@ from .mydesi import MyDesiCrawler
 from .nekohouse import NekohouseCrawler
 from .nhentai import NHentaiCrawler
 from .noodle_magazine import NoodleMagazineCrawler
+from .nsfw_xxx import NsfwXXXCrawler
 from .nudostartv import NudoStarTVCrawler
-from .nxxx import NsfwXXXCrawler
 from .odnoklassniki import OdnoklassnikiCrawler
 from .omegascans import OmegaScansCrawler
 from .onedrive import OneDriveCrawler
